services/realtime-log-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-socketio==5.11.2
python-jose[cryptography]==3.3.0
